No floats plz

+++ moq-karp/src/audio/mod.rs
@@ -23,5 +23,5 @@ pub struct Audio {
 	pub sample_rate: u32,
 	pub channel_count: u32,
 
-	pub bitrate: Option<f64>,
+	pub bitrate: Option<u64>,
 }
